fix issue in recipe ingredient selection if 2 foods share the name
Before this commit it was not possible to Select "Apple" in the Recipe Ingredient selection because "Pineapple" also includes the word "apple" and the search result was not distinct.
This commit changes the behaviour so that if a food is selected in the food search result, it is fetched by id and returns a "definite" food.