fix flaky system tests
Certain system tests were flaky due to race conditions.
The togglers (which are only needed when JavaScript is used) in toggleable menus didn't have a "hidden" class on them, that's why they were discoverable and clickable already even though the JS was not ready for the click event yet.

Also, there was a similar issue in the system test for a "confirmable controller".
The confirmable controller, at some point, has a Submit button with text "Delete" and a regular button "Delete". With JS, the regular button is supposed to be the "first" step of "Delete", the second step is the submit button which is being renamed to "Confirm". Due to a race condition, both buttons are clickable with the text "Delete", but then it submits the form already without going through the second step.